+ Fix: Solucion del error que no mostraba el precio total de la lista de platos del pedido, solo mostraba el precio del ultimo plato seleccionado.

+++ negocio/MesasService.cs
@@ -334,7 +334,7 @@ namespace negocio
             }
         }
 
-        public void ActualizarPrecioEnPedido(float total, int idPedido)
+        public void ActualizarPrecioEnPedido(decimal total, int idPedido)
         {
             try
             {
